Move text extraction to its own transformer.
